+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.5.0 (unreleased)
+* Use `ffmpeg` to generate audio fingerprints instead of `fpcalc`
+  * Requires that the installed version of `ffmpeg`:
+    * Was compiled with the `--enable-chromaprint` option
+    * Understands the `-fp_format raw` flag
+  * `jellyfin-ffmpeg 5.0.1-5` meets both of these requirements
+
 ## v0.1.0.0 (2022-06-09)
 * Add option to automatically skip intros
 * Cache audio fingerprints by default